batch_inference/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
minio>=7.1.0
pyarrow>=7.0.0
mlflow>=2.8.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0 